% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_scale}
\alias{add_scale}
\title{Colour scale.}
\usage{
add_scale(legbreaks, leglabels, col, xlim, ylim, cex = par("cex"))
}
\arguments{
\item{legbreaks}{Set of legend breaks.}

\item{leglabels}{Set of legend labels.}

\item{col}{Set of colours.}

\item{xlim}{X min, max for placement.}

\item{ylim}{Y min, max for placement.}

\item{cex}{Text size (optional).}
}
\description{
Add colour scale to a figure. Adapted from code by Jonty Rougier.
}
