% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_emulation.R
\name{emulate_era}
\alias{emulate_era}
\title{Gaussian process emulation.}
\usage{
emulate_era(sim_inputs, sim_outputs, myvar, trend, covtype,
  expt_design_par)
}
\arguments{
\item{sim_inputs}{Simulator ensemble inputs, i.e. parameter values.}

\item{sim_outputs}{Simulator ensemble outputs, i.e. variable values.}

\item{myvar}{Variable to emulate.}

\item{trend}{Gaussian Process mean function.}

\item{covtype}{Gaussian Process covariance function.}

\item{expt_design_par}{Parameter values at which to make predictions.}
}
\description{
Calibrate GP model with DeConto and Pollard simulator data and predict values at large set of new
parameter values. Uses DiceKriging package.
}
