% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main steering function.}
\usage{
main(analysis_type)
}
\arguments{
\item{analysis_type}{Analysis type: MICI, NoMICI or SimHigh.}
}
\description{
This is a test paragraph.
}
\details{
These are the details.
}
