% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_params.R
\name{plot_params}
\alias{plot_params}
\title{Plot RCP8.5 at 2100 versus 3 parameters.}
\usage{
plot_params(var_plot, param_list, sim_inputs, sim_outputs, em_data)
}
\arguments{
\item{var_plot}{Variable to plot (deprecated: only uses RCP8.5 at 2100).}

\item{param_list}{List of parameters (deprecated: VCLIF, CREVLIQ, OCFAC).}

\item{sim_inputs}{Simulator parameter values.}

\item{sim_outputs}{Simulator variable values.}

\item{em_data}{Emulator parameter and variable values.}
}
\description{
Plot RCP8.5 at 2100 versus 3 parameters.
}
