% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sens_pliocene.R
\name{plot_sens_pliocene}
\alias{plot_sens_pliocene}
\title{Plot sensitivity of results to Pliocene minimum.}
\usage{
plot_sens_pliocene(dataset, calib_data, calib_em = NULL, source,
  var_future, dobias = NA)
}
\arguments{
\item{dataset}{Dataset to use: simulated or emulated.}

\item{calib_data}{Calibration data ranges.}

\item{calib_em}{(Optional) Calibration index for emulated ensemble.}

\item{source}{Data type: "Simulated" or "Emulated".}

\item{var_future}{(Deprecated: RCP8.5 at 2100).}

\item{dobias}{Bias on or off (simulated only) or both (used for emulated).}
}
\description{
Plot sensitivity of results to Pliocene minimum.
}
