% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simulated.R
\name{plot_simulated}
\alias{plot_simulated}
\title{Plot DeConto and Pollard (2016) simulator data.}
\usage{
plot_simulated(sim_data, calib_sim, pliocene_range, chebyshev)
}
\arguments{
\item{sim_data}{Simulation dataset.}

\item{calib_sim}{Calibration index for simulator ensemble.}

\item{pliocene_range}{Name of Pliocene range (used for plot details).}

\item{chebyshev}{k-values for Chebyshev inequality intervals to output to text file.}
}
\description{
Plot DeConto and Pollard (2016) simulator data.
}
