\name{ritz_dens}
\alias{ritz_dens}
\docType{data}
\title{
Density estimate at 2100 from Ritz et al. (2015).
}
\description{
A 512 point density estimate (x and y values) taken from the Supplementary Information of this paper.
}
\usage{data("ritz_dens")}
\format{
Just a single object returned from density().
}
\source{https://media.nature.com/original/nature-assets/nature/journal/v528/n7580/extref/nature16147-s3.zip
}
\references{
Ritz et al. (2015) Nature (https://www.nature.com/articles/nature16147)
}
\examples{
data(ritz_dens)
}
\keyword{datasets}
