\name{ritz_quant}
\alias{ritz_quant}
\docType{data}
\title{
Quantile estimates at 2100 from Ritz et al. (2015) Nature.
}
\description{
Five quantiles calculated from the empirical CDF in the Supplementary Information of this paper.
}
\usage{data("ritz_quant")}
\format{
  The format is:
List of 5
 $ q0.05: num 1.97
 $ q0.25: num 5.45
 $ q0.5 : num 11.9
 $ q0.75: num 17.9
 $ q0.95: num 29.6
}

\source{
https://media.nature.com/original/nature-assets/nature/journal/v528/n7580/extref/nature16147-s3.zip
}
\references{
Ritz et al. (2015) Nature (https://www.nature.com/articles/nature16147)
}
\examples{
data(ritz_quant)
}
\keyword{datasets}
