% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_scale}
\alias{add_scale}
\title{Add colour scale to figure.}
\usage{
add_scale(legbreaks, leglabels, col, xlim, ylim, cex = par("cex"))
}
\arguments{
\item{legbreaks}{Legend breaks.}

\item{leglabels}{Legend labels.}

\item{col}{Colours.}

\item{xlim}{X range.}

\item{ylim}{Y range.}

\item{cex}{Text size (optional).
add_scale(breaks_SLE, sprintf("%.1f", breaks_SLE), colrng_SLE,
xlim = c(9, 9.5), ylim = c(4, 8), cex = 0.7)}
}
\description{
Add colour scale to figure.
}
