% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_emulated.R
\name{plot_emulated}
\alias{plot_emulated}
\title{Plot and write out emulator results.}
\usage{
plot_emulated(sim_data, em_data, calib_data, calib_sim, calib_em, a_type,
  time, calib_era, cred_list, ep_list)
}
\arguments{
\item{sim_data}{Simulation data for plots.}

\item{em_data}{Emulation data.}

\item{calib_data}{Calibration data ranges.}

\item{calib_sim}{Calibration index for simulator ensemble.}

\item{calib_em}{Calibration index for emulator ensemble.}

\item{a_type}{Analysis type: MICI or NoMICI.}

\item{time}{Year of projections (deprecated: only 2100 are in input csv file).}

\item{calib_era}{Which era(s) to use for calibration (deprecated: only threeEras is tested).}

\item{cred_list}{List of credibility thresholds for data file.}

\item{ep_list}{List of exceedance probability thresholds for data file.}
}
\description{
Plot and write out emulator results.
}
